promote type/implementation-specific persistence logic to helix-view
