+++ helix-view/src/lib.rs
@@ -13,6 +13,7 @@ pub mod handlers;
 pub mod info;
 pub mod input;
 pub mod keyboard;
+pub mod persistence;
 pub mod register;
 pub mod theme;
 pub mod tree;
